Extract out penguin entity, and an animation event forwarding script
