--- Assets/PenguinQuest/Code/Controllers/AlwaysOnComponents/PenguinSkeleton.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c415f633318d75347abc38dc748c8bc5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
