+++ Assets/PenguinQuest/Code/Controllers/Handlers/HorizontalMoveHandler.cs
@@ -6,8 +6,7 @@ using PenguinQuest.Data;
 // todo: use distances, max speed, and transition percent per step for easier inspector usage..
 namespace PenguinQuest.Controllers
 {
-    [RequireComponent(typeof(Animator))]
-    [RequireComponent(typeof(Rigidbody2D))]
+    [RequireComponent(typeof(PenguinEntity))]
     public class HorizontalMoveHandler : MonoBehaviour
     {
         private enum Facing { Left = -1, Right = 1}
@@ -24,19 +23,18 @@ namespace PenguinQuest.Controllers
         [Range(0.50f, 100.00f)] [SerializeField] private float maxInputSpeed = 10.0f;
 
 
-        private Rigidbody2D penguinRigidbody;
-        private Animator    penguinAnimator;
-        
+        private PenguinEntity penguinEntity;
+
         private bool   isHorizontalInputActive;
         private float  xMotionIntensity;
         private Facing facing;
         
         void Awake()
         {
-            penguinAnimator  = gameObject.GetComponent<Animator>();
-            penguinRigidbody = gameObject.GetComponent<Rigidbody2D>();
+            penguinEntity = gameObject.GetComponent<PenguinEntity>();
+
             xMotionIntensity = 0.00f;
-            facing           = GetFacing(penguinRigidbody);
+            facing           = GetFacing(penguinEntity.Rigidbody);
         }
         
         void Update()
@@ -61,7 +59,8 @@ namespace PenguinQuest.Controllers
             {
                 //MoveHorizontal(penguinRigidbody, xMotionIntensity * maxInputSpeed, Time.deltaTime);
             }
-            penguinAnimator.SetFloat("XMotionIntensity", xMotionIntensity);
+
+            penguinEntity.Animation.SetParamXMotionIntensity(xMotionIntensity);
         }
 
         void OnEnable()
