+++ Assets/PenguinQuest/Code/Controllers/PenguinAnimation.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d843ab528b90ca24986ff473f5abc37e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
