+++ Assets/PenguinQuest/Sprites/Characters/Penguin/Penguin_Onbelly_Sliding.anim
@@ -6,7 +6,7 @@ AnimationClip:
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_Name: Penguin_Onbelly_Sliding
+  m_Name: Penguin_OnBelly_Sliding
   serializedVersion: 6
   m_Legacy: 0
   m_Compressed: 0
